fix: block 0 datetime
